Extract substrate lifecycle CLI commands
